Former-commit-id: db6378d844a3d4fd08c861d5907ffc3f67f617e2
Former-commit-id: 75c337fb6cd61370f1c2a57c106c37f351ee9d23 [formerly af30f51942d754c8e3e5635d711f9b542b2df95f] [formerly cf18af50096c1126a4a5c2df41bcb69f8aceded9 [formerly 4b21e2c4eb54cf1771b94511e825a4a262e842fe]] [formerly 671d81cb64fe196126270ce81a79365ce9ed50e0 [formerly 117a908570a7a5d7067a27d158995369a1d1f9ed] [formerly 1e15d076bd10f2bad6bcb6e7972cb7f04966e07e [formerly 93f9bcab9bdd03bb350fe064d5f1ecc536ec1940]]]
Former-commit-id: 891775159227a58da8a8d232cb73a1434877d9d5 [formerly 17166b7df4639183291f0eff4c96da528ec3b19f] [formerly 497ac1af55bcbe79982a4a28fcd814bcd5a7d8af [formerly f42d680bcd6e442ad559af14ad23bce03a625400]]
Former-commit-id: d275a3eedc6b31799215d5b63eb6a44e06d06bcc [formerly f50a407b9f036ad961f93bf1e876e4ad72f8f762]
Former-commit-id: bbf239243bc45ef615d09f95d1c5f5693ffa9b16

+++ code/lammps5Nov16/src/USER-NUFEB/fix_bio_kinetics_ph.cpp
@@ -117,7 +117,7 @@ void FixKineticsPH::solve_ph() {
 
   for (int i = 0; i < kinetics->bgrids; i++) {
     double a = 1e-14;
-    int b = 1;
+    double b = 1;
     int ipH = 1;
 
     double fun;
@@ -222,6 +222,7 @@ void FixKineticsPH::solve_ph() {
           while (ipH < maxIter) {
             gSh = (fb * a - fa * b) / (fb - fa);
             sumActivity = 0.0;
+
             for (int k = 1; k < nnus + 1; k++) {
               denm[k] = (1 + kEq[k][0] / w) * gSh * gSh * gSh + kEq[k][1] * gSh * gSh + kEq[k][2] * kEq[k][1] * gSh
                   + kEq[k][3] * kEq[k][2] * kEq[k][1];
